the feature is complete. needs testing now
