+++ src/LayoutContext.ts
@@ -1,0 +1,6 @@
+import * as React from 'react';
+import defaultSettings from './defaultSettings';
+
+const LayoutContext = React.createContext(defaultSettings);
+
+export default LayoutContext;
